+++ symmetria/__init__.py
@@ -1,13 +1,13 @@
 from symmetria.elements.cycle import Cycle
 from symmetria.elements.permutation import Permutation
 from symmetria.generators.random.api import random, random_generator
-from symmetria.generators.algorithm.api import generate
+from symmetria.generators.algorithm.api import generate, permutation_generator
 from symmetria.elements.cycle_decomposition import CycleDecomposition
 
 __version__ = "0.3.1"
 __all__ = [
     "__version__",
-    "generate",
+    "permutation_generator",
     "random",
     "random_generator",
     "Permutation",