+++ tests/tests_generators/test_api.py
@@ -1,6 +1,6 @@
 import pytest
 
-import symmetria
+from symmetria import permutation_generator
 from tests.test_utils import _check_values
 from tests.tests_generators.test_cases import (
     TEST_SJT_GENERATOR,
@@ -17,7 +17,7 @@ from tests.tests_generators.test_cases import (
 )
 def test_permutation_generator_exceptions(algorithm, degree, error, msg) -> None:
     with pytest.raises(error, match=msg):
-        _ = symmetria.generate(algorithm=algorithm, degree=degree)
+        _ = permutation_generator(algorithm=algorithm, degree=degree)
 
 
 @pytest.mark.parametrize(
@@ -28,7 +28,7 @@ def test_permutation_generator_exceptions(algorithm, degree, error, msg) -> None
 def test_lexicographic_generator(degree, expected_value) -> None:
     _check_values(
         expression=f"symmetria.generate('lexicographic', {degree})",
-        evaluation=list(symmetria.generate(algorithm="lexicographic", degree=degree)),
+        evaluation=list(permutation_generator(algorithm="lexicographic", degree=degree)),
         expected=expected_value,
     )
 
@@ -41,7 +41,7 @@ def test_lexicographic_generator(degree, expected_value) -> None:
 def test_heap_generator(degree, expected_value) -> None:
     _check_values(
         expression=f"symmetria.generate('heap', {degree})",
-        evaluation=list(symmetria.generate(algorithm="heap", degree=degree)),
+        evaluation=list(permutation_generator(algorithm="heap", degree=degree)),
         expected=expected_value,
     )
 
@@ -54,6 +54,6 @@ def test_heap_generator(degree, expected_value) -> None:
 def test_steinhaus_johnson_trotter_generator(degree, expected_value) -> None:
     _check_values(
         expression=f"symmetria.generate('steinhaus-johnson-trotter', {degree})",
-        evaluation=list(symmetria.generate(algorithm="steinhaus-johnson-trotter", degree=degree)),
+        evaluation=list(permutation_generator(algorithm="steinhaus-johnson-trotter", degree=degree)),
         expected=expected_value,
     )